Refactor: Modularize ultimate data extractor into v4 package structure
