+++ ultimate_data_extractor_v4/models/query_models.py
@@ -1,0 +1,36 @@
+from dataclasses import dataclass
+from enum import Enum
+from typing import List, Optional
+
+
+class SearchDepth(Enum):
+    BASIC = "basic"
+    STANDARD = "standard"
+    EXTENSIVE = "extensive"
+    EXHAUSTIVE = "exhaustive"
+
+
+@dataclass
+class QueryInput:
+    query: str
+    entity_context: Optional[str] = None
+    region_context: Optional[str] = None
+    metric_context: Optional[str] = None
+    additional_terms: Optional[List[str]] = None
+    date_from: Optional[str] = None
+    date_to: Optional[str] = None
+    languages: Optional[List[str]] = None
+    include_pdfs: bool = True
+    include_excel: bool = True
+    include_csv: bool = True
+    include_apis: bool = True
+    exclude_regions: Optional[List[str]] = None
+    specific_countries: Optional[List[str]] = None
+    max_sources: int = 50
+    search_depth: SearchDepth = SearchDepth.EXTENSIVE
+    require_peer_reviewed: bool = False
+    min_source_authority: float = 0.3
+    max_processing_time: int = 600
+    enable_caching: bool = True
+    parallel_extraction: bool = True
+
