+++ ultimate_pure_extractor.py
@@ -1,1552 +1,7 @@
 """
-ULTIMATE PURE UNIVERSAL DATA EXTRACTOR
-=====================================================
-Complete production-ready data extraction system with advanced capabilities:
-- Zero data manipulation & hallucination prevention
-- Multi-format extraction (PDF, Excel, CSV, JSON, APIs)
-- 75+ sources with intelligent prioritization (trimmed in this demo)
-- Real-time data & compliance tracking
-- Domain-specific extractors
-- Enterprise-grade performance & security
-
-Version: 3.0 Ultimate (flow-corrected)
-Author: AI Assistant
-License: MIT
-
-Environment Variables (optional):
-    GROQ_API_KEY
-    ALPHA_VANTAGE_KEY
-    FRED_API_KEY
+Compatibility shim for the modular v4 package layout.
+This file forwards the public API to ultimate_data_extractor_v4.
 """
 
-import asyncio
-import json
-import logging
-import os
-import random
-import re
-import statistics
-import time
-from collections import defaultdict
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta
-from enum import Enum
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Set, Union
-from urllib.parse import quote, urlparse
-
-# -----------------------------------------------------------------------------
-# Optional dependencies with graceful fallbacks
-# -----------------------------------------------------------------------------
-
-PDF_AVAILABLE = False
-EXCEL_AVAILABLE = False
-DOCX_AVAILABLE = False
-CRAWLING_AVAILABLE = False
-GROQ_AVAILABLE = False
-
-try:
-    import aiohttp  # type: ignore
-except Exception:  # pragma: no cover
-    aiohttp = None  # type: ignore
-
-try:
-    from bs4 import BeautifulSoup  # type: ignore
-except Exception:  # pragma: no cover
-    BeautifulSoup = None  # type: ignore
-
-try:
-    import pandas as pd  # type: ignore
-except Exception:  # pragma: no cover
-    pd = None  # type: ignore
-
-try:
-    import numpy as np  # noqa: F401  # type: ignore
-except Exception:
-    pass
-
-try:
-    import pdfplumber  # type: ignore
-    PDF_AVAILABLE = True
-except Exception:
-    pass
-
-try:
-    from openpyxl import load_workbook  # type: ignore
-    EXCEL_AVAILABLE = True
-except Exception:
-    pass
-
-try:
-    from docx import Document  # noqa: F401  # type: ignore
-    DOCX_AVAILABLE = True
-except Exception:
-    pass
-
-try:  # Optional advanced crawler (not required)
-    from crawl4ai import AsyncWebCrawler, CrawlerRunConfig, CacheMode  # type: ignore
-    CRAWLING_AVAILABLE = True
-except Exception:
-    pass
-
-try:
-    from groq import Groq  # type: ignore
-    GROQ_AVAILABLE = True
-except Exception:
-    Groq = None  # type: ignore
-
-
-# -----------------------------------------------------------------------------
-# Logging
-# -----------------------------------------------------------------------------
-
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-    handlers=[logging.StreamHandler()]
-)
-logger = logging.getLogger("ultimate_extractor")
-
-
-# -----------------------------------------------------------------------------
-# Enums and Data Models
-# -----------------------------------------------------------------------------
-
-
-class SourcePriority(Enum):
-    CRITICAL = 0
-    HIGH = 1
-    MEDIUM = 2
-    LOW = 3
-
-
-class ContentType(Enum):
-    HTML = "html"
-    PDF = "pdf"
-    EXCEL = "excel"
-    CSV = "csv"
-    JSON = "json"
-    XML = "xml"
-    DOCX = "docx"
-    API = "api"
-
-
-class DataFreshness(Enum):
-    REAL_TIME = "real_time"
-    DAILY = "daily"
-    WEEKLY = "weekly"
-    MONTHLY = "monthly"
-    QUARTERLY = "quarterly"
-    ANNUALLY = "annually"
-    STATIC = "static"
-
-
-@dataclass
-class EnhancedDataSource:
-    name: str
-    base_url: str
-    search_patterns: List[str]
-    priority: SourcePriority
-    authority: float
-    content_types: List[ContentType]
-    data_freshness: DataFreshness
-    rate_limit: float = 2.0
-    requires_auth: bool = False
-    api_key_env: Optional[str] = None
-    geographic_coverage: Optional[List[str]] = None
-    domain_expertise: Optional[List[str]] = None
-    peer_reviewed: bool = False
-    institutional_backing: bool = False
-    copyright_status: str = "fair_use"
-    usage_rights: str = "research_educational"
-    terms_of_service_url: Optional[str] = None
-
-
-@dataclass
-class EnhancedQueryInput:
-    query: str
-    entity_context: Optional[str] = None
-    region_context: Optional[str] = None
-    metric_context: Optional[str] = None
-    additional_terms: Optional[List[str]] = None
-    date_from: Optional[str] = None
-    date_to: Optional[str] = None
-    languages: Optional[List[str]] = None
-    include_pdfs: bool = True
-    include_excel: bool = True
-    include_csv: bool = True
-    include_apis: bool = True
-    exclude_regions: Optional[List[str]] = None
-    specific_countries: Optional[List[str]] = None
-    max_sources: int = 50
-    search_depth: str = "extensive"
-    require_peer_reviewed: bool = False
-    min_source_authority: float = 0.3
-    max_processing_time: int = 600
-    enable_caching: bool = True
-    parallel_extraction: bool = True
-
-
-@dataclass
-class EnhancedRawDataPoint:
-    value: Any
-    context: str
-    source_url: str
-    source_name: str
-    extraction_confidence: float
-    raw_text: str
-    publication_date: Optional[datetime] = None
-    data_freshness_days: Optional[int] = None
-    source_credibility: Optional[float] = None
-    peer_reviewed: Optional[bool] = None
-    is_partial_data: Optional[bool] = None
-    methodology_notes: Optional[str] = None
-    content_type: Optional[ContentType] = None
-    page_number: Optional[int] = None
-    table_reference: Optional[str] = None
-    copyright_status: Optional[str] = None
-    usage_rights: Optional[str] = None
-    required_attribution: Optional[str] = None
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class EnhancedSearchResult:
-    query: str
-    source: str
-    url: str
-    title: str
-    content: str
-    data_points: List[EnhancedRawDataPoint]
-    content_type: ContentType
-    errors_encountered: List[str] = field(default_factory=list)
-    blocked_by_paywall: bool = False
-    rate_limited: bool = False
-    requires_authentication: bool = False
-    content_language: Optional[str] = None
-    extraction_accuracy: Optional[float] = None
-    content_completeness: Optional[float] = None
-    source_authority: Optional[float] = None
-    search_timestamp: datetime = field(default_factory=datetime.now)
-    content_hash: str = ""
-
-    def __post_init__(self):
-        if not self.content_hash:
-            self.content_hash = str(abs(hash(self.content)))[:12]
-
-
-# -----------------------------------------------------------------------------
-# Utilities
-# -----------------------------------------------------------------------------
-
-
-class AntiHallucinationValidator:
-    @staticmethod
-    def verify_data_existence(data_point: EnhancedRawDataPoint, original_content: str) -> float:
-        try:
-            raw_text = (data_point.raw_text or "").lower()
-            content_lower = (original_content or "").lower()
-            if raw_text and raw_text in content_lower:
-                return 1.0
-            value_str = str(data_point.value).lower()
-            if value_str and value_str in content_lower:
-                return 0.9
-            context_words = (data_point.context or "").lower().split()
-            if not context_words:
-                return 0.0
-            matches = sum(1 for w in context_words if w in content_lower)
-            return min(matches / max(len(context_words), 1), 0.8)
-        except Exception:
-            return 0.0
-
-
-class LegalComplianceChecker:
-    @staticmethod
-    def check_compliance(source: EnhancedDataSource, content_length: int) -> Dict[str, Any]:
-        info: Dict[str, Any] = {"compliant": True, "warnings": [], "required_actions": []}
-        if content_length > 10000:
-            info["warnings"].append("Large content extraction - verify fair use")
-        if source.copyright_status not in ["public_domain", "fair_use", "creative_commons"]:
-            info["warnings"].append("Verify copyright permissions")
-        if "commercial" in source.usage_rights and source.usage_rights != "commercial_allowed":
-            info["warnings"].append("Non-commercial use only")
-        if source.institutional_backing:
-            info["required_actions"].append(f"Cite: {source.name}")
-        return info
-
-
-class CacheManager:
-    def __init__(self, cache_dir: str = "cache"):
-        self.cache_dir = Path(cache_dir)
-        self.cache_dir.mkdir(exist_ok=True)
-        self.cache_stats = {"hits": 0, "misses": 0}
-
-    def get_cache_key(self, url: str, query: str) -> str:
-        return str(abs(hash(f"{url}|{query}")))
-
-    def get(self, cache_key: str, ttl_hours: int = 24) -> Optional[Any]:
-        cache_file = self.cache_dir / f"{cache_key}.json"
-        if cache_file.exists():
-            try:
-                with open(cache_file, "r", encoding="utf-8") as f:
-                    data = json.load(f)
-                ts = datetime.fromisoformat(data["timestamp"])  # type: ignore
-                if datetime.now() - ts < timedelta(hours=ttl_hours):
-                    self.cache_stats["hits"] += 1
-                    return data.get("data")
-            except Exception:
-                pass
-        self.cache_stats["misses"] += 1
-        return None
-
-    def set(self, cache_key: str, data: Any) -> None:
-        try:
-            with open(self.cache_dir / f"{cache_key}.json", "w", encoding="utf-8") as f:
-                json.dump({"timestamp": datetime.now().isoformat(), "data": data}, f, default=str)
-        except Exception as e:
-            logger.warning(f"Cache write failed: {e}")
-
-
-class RateLimitManager:
-    def __init__(self):
-        self.last_request: Dict[str, float] = {}
-
-    async def wait_if_needed(self, source_name: str, rate_limit: float) -> None:
-        now = time.time()
-        last = self.last_request.get(source_name)
-        if last is not None:
-            elapsed = now - last
-            if elapsed < rate_limit:
-                await asyncio.sleep(rate_limit - elapsed)
-        self.last_request[source_name] = time.time()
-
-
-class PerformanceTracker:
-    def __init__(self):
-        self.metrics: Dict[str, Any] = {
-            "extraction_accuracy": [],
-            "response_times": [],
-            "total_requests": 0,
-            "successful_extractions": 0,
-            "error_rates": defaultdict(int),
-        }
-
-    def record_extraction(self, accuracy: float, response_time: float, source: str, success: bool = True) -> None:
-        self.metrics["extraction_accuracy"].append(accuracy)
-        self.metrics["response_times"].append(response_time)
-        self.metrics["total_requests"] += 1
-        if success:
-            self.metrics["successful_extractions"] += 1
-        else:
-            self.metrics["error_rates"][source] += 1
-
-    def get_summary(self) -> Dict[str, Any]:
-        return {
-            "avg_accuracy": statistics.mean(self.metrics["extraction_accuracy"]) if self.metrics["extraction_accuracy"] else 0,
-            "avg_response_time": statistics.mean(self.metrics["response_times"]) if self.metrics["response_times"] else 0,
-            "success_rate": self.metrics["successful_extractions"] / max(self.metrics["total_requests"], 1),
-            "total_extractions": self.metrics["total_requests"],
-            "error_summary": dict(self.metrics["error_rates"]),
-        }
-
-
-# -----------------------------------------------------------------------------
-# Sources Registry (trimmed set, extensible)
-# -----------------------------------------------------------------------------
-
-
-class ComprehensiveSourceRegistry:
-    def __init__(self):
-        self.sources = self._initialize_sources()
-        logger.info(f"Initialized {len(self.sources)} data sources")
-
-    def _initialize_sources(self) -> List[EnhancedDataSource]:
-        return [
-            EnhancedDataSource(
-                name="World Bank Open Data",
-                base_url="https://data.worldbank.org",
-                search_patterns=["https://data.worldbank.org/search?q={query}"],
-                priority=SourcePriority.CRITICAL,
-                authority=0.98,
-                content_types=[ContentType.HTML, ContentType.API],
-                data_freshness=DataFreshness.QUARTERLY,
-                requires_auth=False,
-                geographic_coverage=["Global"],
-                domain_expertise=["development", "economics"],
-                peer_reviewed=True,
-                institutional_backing=True,
-            ),
-            EnhancedDataSource(
-                name="International Energy Agency",
-                base_url="https://www.iea.org",
-                search_patterns=["https://www.iea.org/search?q={query}"],
-                priority=SourcePriority.HIGH,
-                authority=0.97,
-                content_types=[ContentType.HTML, ContentType.EXCEL],
-                data_freshness=DataFreshness.MONTHLY,
-                geographic_coverage=["Global"],
-                domain_expertise=["energy", "statistics"],
-                peer_reviewed=True,
-                institutional_backing=True,
-            ),
-            EnhancedDataSource(
-                name="Our World in Data",
-                base_url="https://ourworldindata.org",
-                search_patterns=["https://ourworldindata.org/search?q={query}"],
-                priority=SourcePriority.MEDIUM,
-                authority=0.87,
-                content_types=[ContentType.HTML, ContentType.CSV],
-                data_freshness=DataFreshness.MONTHLY,
-                geographic_coverage=["Global"],
-                domain_expertise=["development data"],
-            ),
-        ]
-
-    def get_sources_by_priority(self, query: EnhancedQueryInput) -> List[EnhancedDataSource]:
-        filtered: List[EnhancedDataSource] = []
-        for source in self.sources:
-            if query.min_source_authority and source.authority < query.min_source_authority:
-                continue
-            if query.require_peer_reviewed and not source.peer_reviewed:
-                continue
-            if query.specific_countries and source.geographic_coverage:
-                if not any(c in source.geographic_coverage for c in query.specific_countries):
-                    continue
-            if query.exclude_regions and source.geographic_coverage:
-                if any(r in source.geographic_coverage for r in query.exclude_regions):
-                    continue
-            filtered.append(source)
-        return sorted(filtered, key=lambda s: (s.priority.value, -s.authority))[: query.max_sources]
-
-
-# -----------------------------------------------------------------------------
-# Crawler
-# -----------------------------------------------------------------------------
-
-
-class UltimateWebCrawler:
-    def __init__(self, max_concurrent: int = 16):
-        self.max_concurrent = max_concurrent
-        self.session: Optional[aiohttp.ClientSession] = None if aiohttp else None
-        self.rate_limiter = RateLimitManager()
-        self.user_agents = [
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-            "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-        ]
-
-    async def __aenter__(self):
-        if aiohttp:
-            connector = aiohttp.TCPConnector(limit=self.max_concurrent, limit_per_host=8)
-            timeout = aiohttp.ClientTimeout(total=60, connect=20)
-            self.session = aiohttp.ClientSession(connector=connector, timeout=timeout)
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb):
-        if self.session:
-            await self.session.close()
-
-    async def fetch_content(self, url: str, source: EnhancedDataSource) -> Optional[Dict[str, Any]]:
-        await self.rate_limiter.wait_if_needed(source.name, source.rate_limit)
-        content_type = self._detect_content_type_from_url(url)
-
-        if CRAWLING_AVAILABLE and content_type == ContentType.HTML:
-            try:
-                config = CrawlerRunConfig(
-                    word_count_threshold=100,
-                    extraction_strategy="NoExtractionStrategy",
-                    page_timeout=45000,
-                    delay_before_return_html=1.5,
-                    cache_mode=CacheMode.BYPASS,
-                    headers={"User-Agent": random.choice(self.user_agents)},
-                )
-                async with AsyncWebCrawler(verbose=False) as crawler:  # type: ignore
-                    result = await crawler.arun(url=url, config=config)
-                    if result.success and (result.markdown or result.cleaned_html or result.html):
-                        html = result.cleaned_html or result.html or result.markdown
-                        return {
-                            "content": self._extract_text_from_html(html or ""),
-                            "title": self._extract_title_from_html(html or ""),
-                            "content_type": ContentType.HTML,
-                            "success": True,
-                        }
-            except Exception as e:
-                logger.debug(f"Advanced crawler failed: {e}")
-
-        if not self.session or not aiohttp:
-            return {"content": "", "success": False, "error": "HTTP client unavailable"}
-
-        try:
-            headers = {
-                "User-Agent": random.choice(self.user_agents),
-                "Accept": self._get_accept_header(content_type),
-                "Accept-Language": "en-US,en;q=0.9",
-            }
-            async with self.session.get(url, headers=headers) as resp:
-                if resp.status != 200:
-                    return {"content": "", "success": False, "error": f"HTTP {resp.status}"}
-
-                if content_type == ContentType.JSON or "application/json" in (resp.headers.get("Content-Type") or ""):
-                    text = await resp.text()
-                    return {"content": text, "content_type": ContentType.JSON, "success": True}
-
-                if content_type in (ContentType.PDF, ContentType.EXCEL, ContentType.CSV):
-                    data = await resp.read()
-                    return {"content": data, "content_type": content_type, "success": True, "binary": True}
-
-                html = await resp.text()
-                return {
-                    "content": self._extract_text_from_html(html),
-                    "title": self._extract_title_from_html(html),
-                    "content_type": ContentType.HTML,
-                    "success": True,
-                }
-        except asyncio.TimeoutError:
-            return {"content": "", "success": False, "error": "Timeout"}
-        except Exception as e:
-            return {"content": "", "success": False, "error": str(e)}
-
-    @staticmethod
-    def _detect_content_type_from_url(url: str) -> ContentType:
-        u = url.lower()
-        if u.endswith(".pdf"):
-            return ContentType.PDF
-        if u.endswith((".xlsx", ".xls")):
-            return ContentType.EXCEL
-        if u.endswith(".csv"):
-            return ContentType.CSV
-        if u.endswith(".json"):
-            return ContentType.JSON
-        if "api" in u and ("json" in u or "data" in u):
-            return ContentType.API
-        return ContentType.HTML
-
-    @staticmethod
-    def _get_accept_header(content_type: ContentType) -> str:
-        if content_type == ContentType.JSON:
-            return "application/json"
-        if content_type == ContentType.CSV:
-            return "text/csv"
-        if content_type == ContentType.XML:
-            return "application/xml"
-        return "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8"
-
-    @staticmethod
-    def _extract_text_from_html(html: str) -> str:
-        if not BeautifulSoup:
-            return re.sub(r"<[^>]+>", " ", html)
-        soup = BeautifulSoup(html, "html.parser")
-        for tag in soup([
-            "script",
-            "style",
-            "nav",
-            "footer",
-            "header",
-            "aside",
-            "menu",
-            "form",
-            "input",
-            "button",
-            "iframe",
-            "noscript",
-            "meta",
-            "link",
-        ]):
-            try:
-                tag.decompose()
-            except Exception:
-                pass
-        text = soup.get_text(separator=" ", strip=True)
-        lines: List[str] = []
-        for line in text.splitlines():
-            cleaned = re.sub(r"\s+", " ", line.strip())
-            if cleaned and len(cleaned) > 3:
-                lines.append(cleaned)
-        return "\n".join(lines)
-
-    @staticmethod
-    def _extract_title_from_html(html: str) -> str:
-        if not BeautifulSoup:
-            return "Extracted Content"
-        soup = BeautifulSoup(html, "html.parser")
-        title_tag = soup.find("title")
-        if title_tag:
-            return title_tag.get_text().strip()
-        h1_tag = soup.find("h1")
-        if h1_tag:
-            return h1_tag.get_text().strip()
-        return "Extracted Content"
-
-
-# -----------------------------------------------------------------------------
-# Data Extractor
-# -----------------------------------------------------------------------------
-
-
-class UltimateDataExtractor:
-    def __init__(self, api_key: str):
-        self.model = "llama3-8b-8192"
-        self.validator = AntiHallucinationValidator()
-        self.client = None
-        if GROQ_AVAILABLE and api_key:
-            try:
-                self.client = Groq(api_key=api_key)  # type: ignore
-            except Exception as e:
-                logger.warning(f"Groq init failed, falling back to regex extractor: {e}")
-
-    async def extract_from_content(self, content: Union[str, bytes], query: EnhancedQueryInput, content_type: ContentType) -> List[EnhancedRawDataPoint]:
-        if content_type == ContentType.HTML:
-            return await self._extract_textual(str(content), query, content_type)
-        if content_type == ContentType.JSON:
-            return await self._extract_textual(str(content), query, content_type)
-        if content_type in (ContentType.CSV, ContentType.PDF, ContentType.EXCEL):
-            # Save bytes to a temporary text representation for demo extraction
-            try:
-                if isinstance(content, bytes):
-                    if content_type == ContentType.CSV and pd is not None:
-                        import io
-                        df = pd.read_csv(io.BytesIO(content))  # type: ignore
-                        return await self._extract_textual(df.to_string(), query, content_type)
-                    # For PDF/EXCEL without optional libs, fallback to simple length check
-                    return await self._extract_textual(f"Binary content {content_type.value} ({len(content)} bytes)", query, content_type)
-                else:
-                    return await self._extract_textual(str(content), query, content_type)
-            except Exception as e:
-                logger.debug(f"Binary extraction fallback failed: {e}")
-                return []
-        return await self._extract_textual(str(content), query, content_type)
-
-    async def _extract_textual(self, text: str, query: EnhancedQueryInput, content_type: ContentType) -> List[EnhancedRawDataPoint]:
-        if self.client:
-            try:
-                prompt = self._build_prompt(text, query, content_type)
-                resp = self.client.chat.completions.create(  # type: ignore
-                    messages=[{"role": "user", "content": prompt}],
-                    model=self.model,
-                    temperature=0.05,
-                    max_tokens=1500,
-                )
-                content_out = resp.choices[0].message.content.strip()  # type: ignore
-                data = self._parse_json_response(content_out)
-                points: List[EnhancedRawDataPoint] = []
-                for p in data.get("data_points", []):
-                    if p.get("value") is None:
-                        continue
-                    point = EnhancedRawDataPoint(
-                        value=p.get("value"),
-                        context=p.get("context", ""),
-                        source_url="",
-                        source_name="",
-                        extraction_confidence=float(p.get("confidence", 0.5)),
-                        raw_text=p.get("raw_text", ""),
-                        content_type=content_type,
-                        methodology_notes=p.get("methodology", ""),
-                        metadata={"extraction_method": "ai_llm"},
-                    )
-                    score = self.validator.verify_data_existence(point, text)
-                    point.extraction_confidence *= score
-                    if point.extraction_confidence > 0.1:
-                        points.append(point)
-                return points
-            except Exception as e:
-                logger.debug(f"AI extraction failed, falling back to regex: {e}")
-        return self._regex_extract(text, content_type)
-
-    def _regex_extract(self, text: str, content_type: ContentType) -> List[EnhancedRawDataPoint]:
-        nums = re.findall(r"[-+]?\d*[\.,]?\d+%?", text)[:5]
-        points: List[EnhancedRawDataPoint] = []
-        for n in nums:
-            points.append(
-                EnhancedRawDataPoint(
-                    value=n,
-                    context=text[:120],
-                    source_url="",
-                    source_name="",
-                    extraction_confidence=0.3,
-                    raw_text=n,
-                    content_type=content_type,
-                    metadata={"extraction_method": "regex_fallback"},
-                )
-            )
-        return points
-
-    def _build_prompt(self, content: str, query: EnhancedQueryInput, content_type: ContentType) -> str:
-        add_terms = ", ".join(query.additional_terms or [])
-        date_constraint = ""
-        if query.date_from or query.date_to:
-            date_constraint = f"\nDATE: {query.date_from or 'any'} to {query.date_to or 'any'}"
-        return (
-            f"EXTRACT RAW NUMERICAL DATA (no synthesis)\n"
-            f"Query: {query.query}\nEntity: {query.entity_context or 'n/a'}\nRegion: {query.region_context or 'n/a'}\n"
-            f"Metric: {query.metric_context or 'n/a'}\nTerms: {add_terms}{date_constraint}\n"
-            f"CONTENT_TYPE: {content_type.value}\n"
-            f"CONTENT:\n{content[:8000]}\n\n"
-            "Return JSON with {'data_points': [{'value','context','confidence','raw_text','methodology'}]}"
-        )
-
-    def _parse_json_response(self, response: str) -> Dict[str, Any]:
-        try:
-            if response.strip().startswith("{"):
-                return json.loads(response)
-        except Exception:
-            pass
-        try:
-            if "```json" in response:
-                s = response.find("```json") + 7
-                e = response.find("```", s)
-                return json.loads(response[s:e].strip())
-        except Exception:
-            pass
-        try:
-            start = response.find("{")
-            if start != -1:
-                brace = 0
-                for i, ch in enumerate(response[start:], start):
-                    if ch == "{":
-                        brace += 1
-                    elif ch == "}":
-                        brace -= 1
-                        if brace == 0:
-                            return json.loads(response[start : i + 1])
-        except Exception:
-            pass
-        return {"data_points": []}
-
-
-# -----------------------------------------------------------------------------
-# Search Engine
-# -----------------------------------------------------------------------------
-
-
-class UltimateSearchEngine:
-    def __init__(self, groq_api_key: str):
-        self.source_registry = ComprehensiveSourceRegistry()
-        self.data_extractor = UltimateDataExtractor(groq_api_key)
-        self.cache_manager = CacheManager()
-        self.performance_tracker = PerformanceTracker()
-        self.compliance_checker = LegalComplianceChecker()
-
-    async def search_ultimate(self, query: EnhancedQueryInput) -> List[EnhancedSearchResult]:
-        sources = self.source_registry.get_sources_by_priority(query)
-        searches_per_source = {"basic": 1, "standard": 2, "extensive": 3, "exhaustive": 4}.get(query.search_depth, 3)
-        all_results: List[EnhancedSearchResult] = []
-        max_concurrent = {"basic": 8, "standard": 12, "extensive": 16, "exhaustive": 20}.get(query.search_depth, 16)
-        semaphore = asyncio.Semaphore(max_concurrent)
-
-        async with UltimateWebCrawler(max_concurrent=max_concurrent) as crawler:
-            tasks: List[asyncio.Task] = []
-            variations = self._generate_intelligent_variations(query)
-            for source in sources:
-                for vq in variations[:searches_per_source]:
-                    tasks.append(asyncio.create_task(self._search_single_source_ultimate(source, vq, query, semaphore, crawler)))
-
-            try:
-                results = await asyncio.wait_for(asyncio.gather(*tasks, return_exceptions=True), timeout=query.max_processing_time)
-            except asyncio.TimeoutError:
-                logger.warning(f"Search timeout after {query.max_processing_time}s")
-                results = []
-
-        for r in results:
-            if isinstance(r, list):
-                all_results.extend(r)
-        processed_results = await self._process_results_ultimate(all_results, query)
-        self.performance_tracker.record_extraction(
-            accuracy=self._calculate_average_confidence(processed_results),
-            response_time=0.0,
-            source="ultimate_search",
-            success=len(processed_results) > 0,
-        )
-        return processed_results
-
-    def _generate_intelligent_variations(self, query: EnhancedQueryInput) -> List[str]:
-        v: List[str] = [query.query]
-        if query.entity_context:
-            v.extend([f"{query.entity_context} {query.query}", f"{query.query} {query.entity_context}"])
-        if query.region_context:
-            v.extend([f"{query.query} {query.region_context}", f"{query.region_context} {query.query}"])
-        if query.metric_context:
-            v.extend([f"{query.metric_context} {query.query}", f"{query.query} {query.metric_context} data"])
-        for term in (query.additional_terms or [])[:3]:
-            v.extend([f"{query.query} {term}", f"{term} {query.query}"])
-        return list(dict.fromkeys(v))
-
-    async def _search_single_source_ultimate(self, source: EnhancedDataSource, search_query: str, query: EnhancedQueryInput, semaphore: asyncio.Semaphore, crawler: UltimateWebCrawler) -> List[EnhancedSearchResult]:
-        async with semaphore:
-            results: List[EnhancedSearchResult] = []
-            for pattern in source.search_patterns:
-                url = pattern.format(query=quote(search_query))
-                try:
-                    cache_key = self.cache_manager.get_cache_key(url, search_query) if query.enable_caching else None
-                    if cache_key:
-                        cached = self.cache_manager.get(cache_key, ttl_hours=self._get_cache_ttl(source))
-                        if cached:
-                            results.extend(cached)
-                            continue
-                    fetch = await crawler.fetch_content(url, source)
-                    if not fetch or not fetch.get("success"):
-                        results.append(
-                            EnhancedSearchResult(
-                                query=search_query,
-                                source=source.name,
-                                url=url,
-                                title="Fetch Failed",
-                                content="",
-                                data_points=[],
-                                content_type=ContentType.HTML,
-                                errors_encountered=[(fetch or {}).get("error", "Unknown error")],
-                                blocked_by_paywall=(fetch or {}).get("blocked_by_paywall", False),
-                                rate_limited=(fetch or {}).get("rate_limited", False),
-                            )
-                        )
-                        continue
-                    content = fetch["content"]
-                    ctype: ContentType = fetch.get("content_type", ContentType.HTML)
-                    if not content:
-                        continue
-                    extracted = await self.data_extractor.extract_from_content(content, query, ctype)
-                    if extracted:
-                        for p in extracted:
-                            p.source_url = url
-                            p.source_name = source.name
-                            p.source_credibility = source.authority
-                            p.peer_reviewed = source.peer_reviewed
-                            if source.data_freshness != DataFreshness.STATIC:
-                                p.data_freshness_days = self._freshness_days(source.data_freshness)
-                        result = EnhancedSearchResult(
-                            query=search_query,
-                            source=source.name,
-                            url=url,
-                            title=fetch.get("title", "Data Source Content"),
-                            content=(content if isinstance(content, str) else f"{ctype.value} bytes: {len(content)}")[:3000],
-                            data_points=extracted,
-                            content_type=ctype,
-                            extraction_accuracy=self._calculate_extraction_accuracy(extracted),
-                            content_completeness=min((len(content) if isinstance(content, str) else len(content)) / 5000, 1.0),
-                            source_authority=source.authority,
-                        )
-                        results.append(result)
-                        if query.enable_caching and cache_key:
-                            self.cache_manager.set(cache_key, [result])
-                except Exception as e:
-                    logger.debug(f"Search failed for {source.name}: {e}")
-                    continue
-            return results
-
-    def _get_cache_ttl(self, source: EnhancedDataSource) -> int:
-        mapping = {
-            DataFreshness.REAL_TIME: 1,
-            DataFreshness.DAILY: 6,
-            DataFreshness.WEEKLY: 24,
-            DataFreshness.MONTHLY: 168,
-            DataFreshness.QUARTERLY: 720,
-            DataFreshness.ANNUALLY: 2160,
-            DataFreshness.STATIC: 8760,
-        }
-        return mapping.get(source.data_freshness, 24)
-
-    @staticmethod
-    def _freshness_days(freshness: DataFreshness) -> int:
-        mapping = {
-            DataFreshness.REAL_TIME: 0,
-            DataFreshness.DAILY: 1,
-            DataFreshness.WEEKLY: 7,
-            DataFreshness.MONTHLY: 30,
-            DataFreshness.QUARTERLY: 90,
-            DataFreshness.ANNUALLY: 365,
-            DataFreshness.STATIC: 999,
-        }
-        return mapping.get(freshness, 30)
-
-    @staticmethod
-    def _calculate_extraction_accuracy(data_points: List[EnhancedRawDataPoint]) -> float:
-        if not data_points:
-            return 0.0
-        return sum(p.extraction_confidence for p in data_points) / len(data_points)
-
-    @staticmethod
-    def _calculate_average_confidence(results: List[EnhancedSearchResult]) -> float:
-        vals: List[float] = []
-        for r in results:
-            for p in r.data_points:
-                vals.append(p.extraction_confidence)
-        return statistics.mean(vals) if vals else 0.0
-
-    async def _process_results_ultimate(self, all_results: List[EnhancedSearchResult], query: EnhancedQueryInput) -> List[EnhancedSearchResult]:
-        if not all_results:
-            return []
-        with_data = [r for r in all_results if r.data_points]
-        if not with_data:
-            return all_results[:10]
-        unique = await self._deduplicate_results_advanced(with_data)
-        filtered = self._apply_quality_filters(unique, query)
-        ranked = self._rank_results_intelligently(filtered, query)
-        diverse = self._ensure_source_diversity_advanced(ranked)
-        return diverse
-
-    async def _deduplicate_results_advanced(self, results: List[EnhancedSearchResult]) -> List[EnhancedSearchResult]:
-        seen: Set[str] = set()
-        unique: List[EnhancedSearchResult] = []
-        for r in results:
-            domain = urlparse(r.url).netloc
-            values = [str(p.value) for p in r.data_points[:5]]
-            signature = f"{domain}_{r.title[:50]}_{'_'.join(values)}"
-            sig_hash = str(abs(hash(signature)))
-            if sig_hash not in seen:
-                seen.add(sig_hash)
-                unique.append(r)
-        return unique
-
-    @staticmethod
-    def _apply_quality_filters(results: List[EnhancedSearchResult], query: EnhancedQueryInput) -> List[EnhancedSearchResult]:
-        out: List[EnhancedSearchResult] = []
-        for r in results:
-            if r.source_authority and r.source_authority < query.min_source_authority:
-                continue
-            if query.require_peer_reviewed and not any(p.peer_reviewed for p in r.data_points):
-                continue
-            if r.data_points:
-                avg_conf = statistics.mean(p.extraction_confidence for p in r.data_points)
-                if avg_conf < 0.2:
-                    continue
-            out.append(r)
-        return out
-
-    @staticmethod
-    def _rank_results_intelligently(results: List[EnhancedSearchResult], query: EnhancedQueryInput) -> List[EnhancedSearchResult]:
-        def score(r: EnhancedSearchResult) -> float:
-            s = 0.0
-            if r.source_authority:
-                s += r.source_authority * 0.30
-            s += min(len(r.data_points) / 10.0, 0.25)
-            if r.extraction_accuracy:
-                s += r.extraction_accuracy * 0.20
-            if r.content_completeness:
-                s += r.content_completeness * 0.15
-            terms = query.query.lower().split()
-            txt = r.content.lower()
-            rel = sum(1 for t in terms if t in txt) / max(len(terms), 1)
-            s += rel * 0.10
-            return min(s, 1.0)
-
-        return sorted(results, key=score, reverse=True)
-
-    @staticmethod
-    def _ensure_source_diversity_advanced(results: List[EnhancedSearchResult], max_per_source: int = 5) -> List[EnhancedSearchResult]:
-        out: List[EnhancedSearchResult] = []
-        counts: Dict[str, int] = defaultdict(int)
-        for r in results:
-            if counts[r.source] < max_per_source:
-                out.append(r)
-                counts[r.source] += 1
-        return out[:50]
-
-
-# -----------------------------------------------------------------------------
-# Aggregator
-# -----------------------------------------------------------------------------
-
-
-class UltimatePureDataAggregator:
-    def __init__(self, groq_api_key: str):
-        if not groq_api_key:
-            raise ValueError("Groq API key is required")
-        self.search_engine = UltimateSearchEngine(groq_api_key)
-        self.performance_tracker = PerformanceTracker()
-        logger.info("Ultimate Pure Data Aggregator initialized")
-
-    async def extract_ultimate_data(self, query: EnhancedQueryInput) -> Dict[str, Any]:
-        start = time.time()
-        try:
-            search_results = await self.search_engine.search_ultimate(query)
-            analysis = await self._analyze_results_comprehensively(search_results)
-            processing_time = time.time() - start
-            output: Dict[str, Any] = {
-                "query": query.query,
-                "entity_context": query.entity_context,
-                "region_context": query.region_context,
-                "metric_context": query.metric_context,
-                "additional_terms": query.additional_terms,
-                "search_configuration": {
-                    "search_depth": query.search_depth,
-                    "max_sources": query.max_sources,
-                    "languages": query.languages,
-                    "date_constraints": {"from": query.date_from, "to": query.date_to},
-                    "content_types": {
-                        "pdfs": query.include_pdfs,
-                        "excel": query.include_excel,
-                        "csv": query.include_csv,
-                        "apis": query.include_apis,
-                    },
-                },
-                "raw_data_points": [
-                    {
-                        "value": p.value,
-                        "context": p.context,
-                        "source_name": p.source_name,
-                        "source_url": p.source_url,
-                        "extraction_confidence": p.extraction_confidence,
-                        "raw_text": p.raw_text,
-                        "content_type": p.content_type.value if p.content_type else None,
-                        "publication_date": p.publication_date.isoformat() if p.publication_date else None,
-                        "data_freshness_days": p.data_freshness_days,
-                        "source_credibility": p.source_credibility,
-                        "peer_reviewed": p.peer_reviewed,
-                        "page_number": p.page_number,
-                        "table_reference": p.table_reference,
-                        "methodology_notes": p.methodology_notes,
-                        "legal_compliance": {
-                            "copyright_status": p.copyright_status,
-                            "usage_rights": p.usage_rights,
-                            "required_attribution": p.required_attribution,
-                        },
-                        "metadata": p.metadata,
-                    }
-                    for r in search_results
-                    for p in r.data_points
-                ],
-                "search_analytics": analysis,
-                "processing_metadata": {
-                    "timestamp": datetime.now().isoformat(),
-                    "processing_time_seconds": round(processing_time, 2),
-                    "groq_model": "llama3-8b-8192",
-                    "aggregator_version": "3.0_ultimate",
-                    "total_sources_attempted": len({r.source for r in search_results}),
-                    "successful_extractions": len([r for r in search_results if r.data_points]),
-                    "total_data_points": sum(len(r.data_points) for r in search_results),
-                    "average_confidence": self._calculate_overall_confidence(search_results),
-                    "extraction_methods_used": list({p.metadata.get("extraction_method", "unknown") for r in search_results for p in r.data_points}),
-                    "content_types_processed": list({r.content_type.value for r in search_results}),
-                    "cache_performance": {
-                        "hits": self.search_engine.cache_manager.cache_stats["hits"],
-                        "misses": self.search_engine.cache_manager.cache_stats["misses"],
-                        "hit_rate": self.search_engine.cache_manager.cache_stats["hits"] / max(sum(self.search_engine.cache_manager.cache_stats.values()), 1),
-                    },
-                },
-                "performance_metrics": self.performance_tracker.get_summary(),
-                "quality_indicators": {
-                    "peer_reviewed_sources": len([r for r in search_results if any(p.peer_reviewed for p in r.data_points)]),
-                    "high_authority_sources": len([r for r in search_results if (r.source_authority or 0) > 0.8]),
-                    "recent_data_points": len([p for r in search_results for p in r.data_points if (p.data_freshness_days or 999) < 30]),
-                    "extraction_accuracy_avg": analysis.get("average_extraction_accuracy", 0),
-                    "source_diversity_score": len({r.source for r in search_results}) / max(len(search_results), 1),
-                },
-                "search_results_details": [
-                    {
-                        "source": r.source,
-                        "url": r.url,
-                        "title": r.title,
-                        "content_type": r.content_type.value,
-                        "data_points_count": len(r.data_points),
-                        "extraction_accuracy": r.extraction_accuracy,
-                        "source_authority": r.source_authority,
-                        "search_timestamp": r.search_timestamp.isoformat(),
-                        "errors_encountered": r.errors_encountered,
-                        "blocked_by_paywall": r.blocked_by_paywall,
-                        "rate_limited": r.rate_limited,
-                        "content_preview": (r.content[:200] + "...") if len(r.content) > 200 else r.content,
-                    }
-                    for r in search_results
-                ],
-            }
-            return output
-        except Exception as e:
-            logger.error(f"Ultimate query failed: {e}")
-            return self._create_error_output(query, str(e))
-
-    def _calculate_overall_confidence(self, search_results: List[EnhancedSearchResult]) -> float:
-        vals: List[float] = []
-        for r in search_results:
-            for p in r.data_points:
-                vals.append(p.extraction_confidence)
-        return statistics.mean(vals) if vals else 0.0
-
-    def _create_error_output(self, query: EnhancedQueryInput, error: str) -> Dict[str, Any]:
-        return {
-            "query": query.query,
-            "entity_context": query.entity_context,
-            "region_context": query.region_context,
-            "error": error,
-            "error_timestamp": datetime.now().isoformat(),
-            "raw_data_points": [],
-            "search_analytics": {"total_results": 0, "analysis": f"Processing failed: {error}"},
-            "processing_metadata": {
-                "timestamp": datetime.now().isoformat(),
-                "processing_time_seconds": 0,
-                "total_sources_attempted": 0,
-                "successful_extractions": 0,
-                "total_data_points": 0,
-            },
-        }
-
-    async def _analyze_results_comprehensively(self, search_results: List[EnhancedSearchResult]) -> Dict[str, Any]:
-        if not search_results:
-            return {"total_results": 0, "analysis": "No results found"}
-        source_breakdown: Dict[str, Dict[str, Any]] = defaultdict(lambda: {"data_points": 0, "avg_confidence": 0, "content_types": set(), "authority": 0})
-        all_conf: List[float] = []
-        content_type_counts: Dict[str, int] = defaultdict(int)
-        error_summary: Dict[str, int] = defaultdict(int)
-        for r in search_results:
-            sb = source_breakdown[r.source]
-            sb["data_points"] += len(r.data_points)
-            sb["authority"] = r.source_authority or 0
-            sb["content_types"].add(r.content_type.value)
-            pcs = [p.extraction_confidence for p in r.data_points]
-            all_conf.extend(pcs)
-            if pcs:
-                sb["avg_confidence"] = statistics.mean(pcs)
-            content_type_counts[r.content_type.value] += 1
-            for e in r.errors_encountered:
-                error_summary[e] += 1
-        for v in source_breakdown.values():
-            v["content_types"] = list(v["content_types"])
-        return {
-            "total_results": len(search_results),
-            "results_with_data": len([r for r in search_results if r.data_points]),
-            "total_data_points": sum(len(r.data_points) for r in search_results),
-            "average_extraction_accuracy": statistics.mean(all_conf) if all_conf else 0,
-            "confidence_distribution": {
-                "min": min(all_conf) if all_conf else 0,
-                "max": max(all_conf) if all_conf else 0,
-                "median": statistics.median(all_conf) if all_conf else 0,
-                "std_dev": statistics.stdev(all_conf) if len(all_conf) > 1 else 0,
-            },
-            "source_breakdown": dict(source_breakdown),
-            "content_type_distribution": dict(content_type_counts),
-            "error_summary": dict(error_summary),
-            "quality_metrics": {
-                "high_confidence_points": len([c for c in all_conf if c > 0.8]),
-                "medium_confidence_points": len([c for c in all_conf if 0.5 <= c <= 0.8]),
-                "low_confidence_points": len([c for c in all_conf if c < 0.5]),
-            },
-        }
-
-    async def batch_extract_ultimate(self, queries: List[EnhancedQueryInput], max_concurrent: int = 3) -> List[Dict[str, Any]]:
-        semaphore = asyncio.Semaphore(max_concurrent)
-
-        async def _run_single(q: EnhancedQueryInput) -> Dict[str, Any]:
-            async with semaphore:
-                return await self.extract_ultimate_data(q)
-
-        start = time.time()
-        tasks = [asyncio.create_task(_run_single(q)) for q in queries]
-        results = await asyncio.gather(*tasks, return_exceptions=True)
-        out: List[Dict[str, Any]] = []
-        for i, r in enumerate(results):
-            if isinstance(r, Exception):
-                logger.error(f"Query {i} failed: {r}")
-                out.append(self._create_error_output(queries[i], str(r)))
-            else:
-                out.append(r)
-        logger.info(f"Batch complete: {len(out)} results in {time.time() - start:.1f}s")
-        return out
-
-
-# -----------------------------------------------------------------------------
-# Export Helpers
-# -----------------------------------------------------------------------------
-
-
-def _analyze_content_types(results: List[Dict[str, Any]]) -> Dict[str, int]:
-    counts: Dict[str, int] = defaultdict(int)
-    for r in results:
-        for p in r.get("raw_data_points", []):
-            counts[p.get("content_type", "unknown")] += 1
-    return dict(counts)
-
-
-def _analyze_source_diversity(results: List[Dict[str, Any]]) -> Dict[str, Any]:
-    all_sources: Set[str] = set()
-    source_counts: Dict[str, int] = defaultdict(int)
-    for r in results:
-        for p in r.get("raw_data_points", []):
-            sn = p.get("source_name")
-            if sn:
-                all_sources.add(sn)
-                source_counts[sn] += 1
-    return {
-        "unique_sources_total": len(all_sources),
-        "most_used_sources": sorted(source_counts.items(), key=lambda x: x[1], reverse=True)[:10],
-        "source_distribution_stats": {
-            "mean_usage": statistics.mean(source_counts.values()) if source_counts else 0,
-            "max_usage": max(source_counts.values()) if source_counts else 0,
-            "sources_used_once": sum(1 for c in source_counts.values() if c == 1),
-        },
-    }
-
-
-def _generate_batch_summary(results: List[Dict[str, Any]]) -> Dict[str, Any]:
-    successful = [r for r in results if "error" not in r]
-    failed = [r for r in results if "error" in r]
-    summary: Dict[str, Any] = {
-        "total_queries": len(results),
-        "successful_queries": len(successful),
-        "failed_queries": len(failed),
-        "success_rate": (len(successful) / len(results)) if results else 0,
-    }
-    if successful:
-        all_points: List[Dict[str, Any]] = []
-        times: List[float] = []
-        sources_attempted: List[int] = []
-        confs: List[float] = []
-        for r in successful:
-            dps = r.get("raw_data_points", [])
-            all_points.extend(dps)
-            pm = r.get("processing_metadata", {})
-            times.append(pm.get("processing_time_seconds", 0))
-            sources_attempted.append(pm.get("total_sources_attempted", 0))
-            confs.extend([p.get("extraction_confidence", 0) for p in dps])
-        summary.update(
-            {
-                "total_data_points_extracted": len(all_points),
-                "average_data_points_per_query": len(all_points) / max(len(successful), 1),
-                "average_processing_time": statistics.mean(times) if times else 0,
-                "average_sources_per_query": statistics.mean(sources_attempted) if sources_attempted else 0,
-                "overall_confidence_stats": (
-                    {
-                        "mean": statistics.mean(confs) if confs else 0,
-                        "median": statistics.median(confs) if confs else 0,
-                        "min": min(confs) if confs else 0,
-                        "max": max(confs) if confs else 0,
-                    }
-                    if confs
-                    else {}
-                ),
-                "content_type_distribution": _analyze_content_types(successful),
-                "source_diversity": _analyze_source_diversity(successful),
-            }
-        )
-    return summary
-
-
-def export_ultimate_data(results: Union[Dict[str, Any], List[Dict[str, Any]]], base_filename: str = "ultimate_data_export") -> Dict[str, str]:
-    if isinstance(results, dict):
-        results = [results]
-    ts = datetime.now().strftime("%Y%m%d_%H%M%S")
-    export_info: Dict[str, str] = {}
-    json_filename = f"{base_filename}_{ts}.json"
-    with open(json_filename, "w", encoding="utf-8") as f:
-        json.dump(
-            {
-                "export_metadata": {
-                    "timestamp": datetime.now().isoformat(),
-                    "export_type": "ultimate_pure_data_extraction",
-                    "aggregator_version": "3.0_ultimate",
-                    "total_queries": len(results),
-                    "total_data_points": sum(len(r.get("raw_data_points", [])) for r in results),
-                    "groq_model": "llama3-8b-8192",
-                    "features_used": [
-                        "multi_format_extraction",
-                        "anti_hallucination_validation",
-                        "legal_compliance_checking",
-                        "advanced_caching",
-                        "performance_tracking",
-                    ],
-                },
-                "processing_summary": _generate_batch_summary(results),
-                "results": results,
-            },
-            f,
-            indent=2,
-            ensure_ascii=False,
-            default=str,
-        )
-    export_info["json"] = json_filename
-
-    csv_rows: List[Dict[str, Any]] = []
-    for r in results:
-        q = r.get("query", "")
-        ent = r.get("entity_context", "")
-        reg = r.get("region_context", "")
-        for p in r.get("raw_data_points", []):
-            csv_rows.append(
-                {
-                    "Query": q,
-                    "Entity_Context": ent,
-                    "Region_Context": reg,
-                    "Raw_Value": p.get("value"),
-                    "Context": p.get("context", ""),
-                    "Source_Name": p.get("source_name", ""),
-                    "Source_URL": p.get("source_url", ""),
-                    "Extraction_Confidence": p.get("extraction_confidence", 0),
-                    "Content_Type": p.get("content_type", ""),
-                    "Raw_Text": p.get("raw_text", ""),
-                    "Source_Credibility": p.get("source_credibility", 0),
-                    "Peer_Reviewed": p.get("peer_reviewed", False),
-                    "Data_Freshness_Days": p.get("data_freshness_days", ""),
-                    "Page_Number": p.get("page_number", ""),
-                    "Table_Reference": p.get("table_reference", ""),
-                    "Methodology_Notes": p.get("methodology_notes", ""),
-                    "Copyright_Status": (p.get("legal_compliance") or {}).get("copyright_status", ""),
-                    "Usage_Rights": (p.get("legal_compliance") or {}).get("usage_rights", ""),
-                    "Required_Attribution": (p.get("legal_compliance") or {}).get("required_attribution", ""),
-                    "Extraction_Method": (p.get("metadata") or {}).get("extraction_method", ""),
-                }
-            )
-    if csv_rows and pd is not None:
-        csv_filename = f"{base_filename}_{ts}.csv"
-        pd.DataFrame(csv_rows).to_csv(csv_filename, index=False, encoding="utf-8")  # type: ignore
-        export_info["csv"] = csv_filename
-
-    citation_filename = f"{base_filename}_citations_{ts}.txt"
-    with open(citation_filename, "w", encoding="utf-8") as f:
-        f.write("ULTIMATE PURE DATA EXTRACTOR - SOURCE CITATIONS\n")
-        f.write("=" * 50 + "\n\n")
-        for i, r in enumerate(results, 1):
-            f.write(f"Query {i}: {r.get('query', '')}\n")
-            f.write("-" * 30 + "\n")
-            seen: Set[str] = set()
-            for p in r.get("raw_data_points", []):
-                sn = p.get("source_name")
-                su = p.get("source_url")
-                if sn and su:
-                    key = f"{sn}|{su}"
-                    if key in seen:
-                        continue
-                    seen.add(key)
-                    f.write(f"- {sn}\n  URL: {su}\n  Accessed: {datetime.now().strftime('%Y-%m-%d')}\n\n")
-            f.write("\n")
-    export_info["citations"] = citation_filename
-    return export_info
-
-
-# -----------------------------------------------------------------------------
-# Public API & CLI helpers
-# -----------------------------------------------------------------------------
-
-
-async def extract_ultimate_pure_data(
-    query_text: str,
-    entity_context: Optional[str] = None,
-    region_context: Optional[str] = None,
-    metric_context: Optional[str] = None,
-    additional_terms: Optional[List[str]] = None,
-    date_from: Optional[str] = None,
-    date_to: Optional[str] = None,
-    search_depth: str = "extensive",
-    max_sources: int = 50,
-    languages: Optional[List[str]] = None,
-    include_pdfs: bool = True,
-    include_excel: bool = True,
-    include_csv: bool = True,
-    include_apis: bool = True,
-    require_peer_reviewed: bool = False,
-    min_source_authority: float = 0.3,
-    specific_countries: Optional[List[str]] = None,
-    exclude_regions: Optional[List[str]] = None,
-    max_processing_time: int = 600,
-    enable_caching: bool = True,
-    groq_api_key: Optional[str] = None,
-) -> Dict[str, Any]:
-    api_key = groq_api_key or os.getenv("GROQ_API_KEY")
-    if not api_key:
-        raise ValueError("Groq API key required")
-    query = EnhancedQueryInput(
-        query=query_text,
-        entity_context=entity_context,
-        region_context=region_context,
-        metric_context=metric_context,
-        additional_terms=additional_terms or [],
-        date_from=date_from,
-        date_to=date_to,
-        search_depth=search_depth,
-        max_sources=max_sources,
-        languages=languages,
-        include_pdfs=include_pdfs,
-        include_excel=include_excel,
-        include_csv=include_csv,
-        include_apis=include_apis,
-        require_peer_reviewed=require_peer_reviewed,
-        min_source_authority=min_source_authority,
-        specific_countries=specific_countries,
-        exclude_regions=exclude_regions,
-        max_processing_time=max_processing_time,
-        enable_caching=enable_caching,
-    )
-    aggregator = UltimatePureDataAggregator(api_key)
-    return await aggregator.extract_ultimate_data(query)
-
-
-async def main_ultimate_demo() -> List[Dict[str, Any]]:
-    api_key = os.getenv("GROQ_API_KEY")
-    if not api_key:
-        print("GROQ_API_KEY not found")
-        return []
-    demo = UltimatePureDataDemo(api_key)
-    return await demo.run_ultimate_demo()
-
-
-async def custom_ultimate_query() -> Optional[Dict[str, Any]]:
-    api_key = os.getenv("GROQ_API_KEY")
-    if not api_key:
-        print("GROQ_API_KEY not found")
-        return None
-    print("\nCUSTOM ULTIMATE QUERY\n" + "=" * 50)
-    query_text = input("Enter your query: ").strip()
-    if not query_text:
-        print("No query provided")
-        return None
-    entity_context = (input("Entity context (optional): ").strip() or None)
-    region_context = (input("Region context (optional): ").strip() or None)
-    metric_context = (input("Metric context (optional): ").strip() or None)
-    additional_terms_input = input("Additional terms (comma-separated, optional): ").strip()
-    additional_terms = [t.strip() for t in additional_terms_input.split(",")] if additional_terms_input else None
-    date_from = (input("Date from (YYYY-MM-DD, optional): ").strip() or None)
-    date_to = (input("Date to (YYYY-MM-DD, optional): ").strip() or None)
-    search_depth = (input("Search depth (basic/standard/extensive/exhaustive) [extensive]: ").strip() or "extensive")
-    max_sources = int(input("Max sources (1-75) [50]: ").strip() or "50")
-    include_pdfs = input("Include PDFs? (y/n) [y]: ").strip().lower() != "n"
-    include_excel = input("Include Excel files? (y/n) [y]: ").strip().lower() != "n"
-    include_apis = input("Include API sources? (y/n) [y]: ").strip().lower() != "n"
-    require_peer_reviewed = input("Require peer-reviewed sources? (y/n) [n]: ").strip().lower() == "y"
-    min_source_authority = float(input("Minimum source authority (0.0-1.0) [0.3]: ").strip() or "0.3")
-    print(f"\nProcessing: {query_text}")
-    try:
-        result = await extract_ultimate_pure_data(
-            query_text=query_text,
-            entity_context=entity_context,
-            region_context=region_context,
-            metric_context=metric_context,
-            additional_terms=additional_terms,
-            date_from=date_from,
-            date_to=date_to,
-            search_depth=search_depth,
-            max_sources=max_sources,
-            include_pdfs=include_pdfs,
-            include_excel=include_excel,
-            include_apis=include_apis,
-            require_peer_reviewed=require_peer_reviewed,
-            min_source_authority=min_source_authority,
-            groq_api_key=api_key,
-        )
-        export_info = export_ultimate_data(result, "custom_ultimate_query")
-        print("\nExported:")
-        for k, v in export_info.items():
-            print(f"  {k.upper()}: {v}")
-        return result
-    except Exception as e:
-        print(f"Ultimate query failed: {e}")
-        return None
-
-
-async def test_ultimate_api() -> None:
-    api_key = os.getenv("GROQ_API_KEY")
-    if not api_key:
-        print("GROQ_API_KEY not set")
-        return
-    tests = [
-        {"name": "Basic", "query": "solar energy capacity", "config": {"search_depth": "basic", "max_sources": 3}},
-        {"name": "Advanced", "query": "renewable energy statistics global trends", "config": {"search_depth": "standard", "max_sources": 5}},
-    ]
-    for i, t in enumerate(tests, 1):
-        print(f"\nRunning {t['name']} ({i}/{len(tests)})...")
-        try:
-            start = time.time()
-            r = await extract_ultimate_pure_data(query_text=t["query"], groq_api_key=api_key, **t["config"])  # type: ignore
-            dt = time.time() - start
-            if r and not r.get("error"):
-                print(f"  PASSED in {dt:.2f}s; points={len(r.get('raw_data_points', []))}")
-            else:
-                print(f"  FAILED: {r.get('error', 'No data')} ")
-        except Exception as e:
-            print(f"  FAILED: {e}")
-
-
-def show_ultimate_input_format() -> None:
-    sample = {
-        "query_text": "renewable energy capacity statistics worldwide",
-        "entity_context": "Renewable Energy",
-        "region_context": "Global",
-        "metric_context": "installed capacity",
-        "additional_terms": ["wind", "solar", "hydro", "biomass"],
-        "date_from": "2020-01-01",
-        "date_to": "2024-12-31",
-        "search_depth": "extensive",
-        "max_sources": 50,
-        "max_processing_time": 600,
-        "include_pdfs": True,
-        "include_excel": True,
-        "include_csv": True,
-        "include_apis": True,
-        "require_peer_reviewed": False,
-        "min_source_authority": 0.3,
-        "specific_countries": ["Germany", "France", "UK"],
-        "exclude_regions": ["Middle East"],
-        "languages": ["en", "es", "fr"],
-        "enable_caching": True,
-        "parallel_extraction": True,
-    }
-    print(json.dumps(sample, indent=2))
-
-
-# -----------------------------------------------------------------------------
-# Demo Wrapper
-# -----------------------------------------------------------------------------
-
-
-class UltimatePureDataDemo:
-    def __init__(self, groq_api_key: str):
-        self.aggregator = UltimatePureDataAggregator(groq_api_key)
-
-    def get_ultimate_sample_queries(self) -> List[EnhancedQueryInput]:
-        return [
-            EnhancedQueryInput(
-                query="global renewable energy capacity statistics",
-                entity_context="Renewable Energy",
-                region_context="Global",
-                metric_context="installed capacity",
-                additional_terms=["wind", "solar", "hydro", "geothermal"],
-                date_from="2020-01-01",
-                date_to="2024-12-31",
-                languages=["en"],
-                search_depth="extensive",
-                max_sources=10,
-            ),
-            EnhancedQueryInput(
-                query="electric vehicle market share data worldwide",
-                entity_context="Electric Vehicles",
-                region_context="Global",
-                metric_context="market penetration",
-                additional_terms=["BEV", "PHEV", "sales volume", "market share"],
-                search_depth="standard",
-                max_sources=8,
-            ),
-        ]
-
-    async def run_ultimate_demo(self) -> List[Dict[str, Any]]:
-        qs = self.get_ultimate_sample_queries()
-        results = await self.aggregator.batch_extract_ultimate(qs, max_concurrent=2)
-        export_info = export_ultimate_data(results, "ultimate_demo")
-        print("Exports:")
-        for k, v in export_info.items():
-            print(f"  {k.upper()}: {v}")
-        return results
-
-
-# -----------------------------------------------------------------------------
-# CLI
-# -----------------------------------------------------------------------------
-
-
-if __name__ == "__main__":
-    print("ULTIMATE PURE UNIVERSAL DATA EXTRACTOR\n" + "=" * 50)
-    print("1. Run Ultimate Demo")
-    print("2. Custom Ultimate Query")
-    print("3. Test Ultimate API")
-    print("4. Show Input Format Guide")
-    try:
-        choice = input("Enter choice (1-4): ").strip()
-        if choice == "1":
-            asyncio.run(main_ultimate_demo())
-        elif choice == "2":
-            asyncio.run(custom_ultimate_query())
-        elif choice == "3":
-            asyncio.run(test_ultimate_api())
-        elif choice == "4":
-            show_ultimate_input_format()
-        else:
-            asyncio.run(main_ultimate_demo())
-    except KeyboardInterrupt:
-        print("\nGoodbye!")
-    except Exception as e:
-        print(f"Error: {e}")
+from ultimate_data_extractor_v4.main import extract_ultimate_pure_data  # noqa: F401
